Big Update/Fixes
Fixed bug with the iteratively generated button.
Implemented ability to switch between expansion through button on the left, as well as adding all the expansion up the Battle for Azeroth and an introductionary page.
Also made sure the probabilty printed to be a precentage and have a precision of four